1-last_digit.c task for alx project

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -14,20 +14,20 @@ int main(void)
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
-	int l = n % 10;
+	long int l = n % 10;
 
 	if (l > 5)
 
 	{
-		printf("Last digit of %d is %d and is greater than 5\n", n, l);
+		printf("Last digit of %d is %ld and is greater than 5\n", n, l);
 	}
 	else if (l < 6 && l != 0)
 	{
-		printf("Last digit of %d is %d and is less than 6 and not 0\n", n, l);
+		printf("Last digit of %d is %ld and is less than 6 and not 0\n", n, l);
 	}
 	else if (l == 0)
 	{
-		printf("Last digit of %d is %d an is 0", n, l);
+		printf("Last digit of %d is %ld an is 0", n, l);
 	}
 	return (0);
 }
